:recycle: Move annotations to delegate; move inter extensions
