+++ libs/designsystem/src/lib/components/chart/configs/chart-config.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ChartType as ChartJSType } from 'chart.js';
+import { ActiveElement, Chart, ChartOptions, ChartType as ChartJSType } from 'chart.js';
 import { AnnotationType, AnnotationTypeRegistry } from 'chartjs-plugin-annotation';
 
 import { deepCopy } from '../../../helpers/deep-copy';
@@ -21,14 +21,21 @@ export class ChartConfigService {
     return CHART_ANNOTATIONS_CONFIG[type];
   }
 
-  public getInteractionFunctionsExtensions() {
-    return CHART_INTERACTION_FUNCTIONS_EXTENSIONS;
-  }
-
   /* Our types does not always map 1 to 1 to the same type 
   that the chart.js chart is actually oconfigured with. Therefore this function 
   for looking up the ChartJSType of a type.  */
   public chartTypeToChartJSType(chartType: ChartType): ChartJSType {
     return CHART_TYPES_CONFIG[chartType]['type'] as ChartJSType;
+  }
+
+  public applyInteractionFunctionsExtensions(options: ChartOptions): ChartOptions {
+    const interactionFunctionsExtensions = CHART_INTERACTION_FUNCTIONS_EXTENSIONS;
+    Object.entries(interactionFunctionsExtensions).forEach(([key, _]) => {
+      const callback = options[key];
+      options[key] = (e: Event, a: ActiveElement[], c: Chart) => {
+        interactionFunctionsExtensions[key](e, a, c, callback);
+      };
+    });
+    return options;
   }
 }
